Fix assert
